% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BucksM.R
\name{buck.im}
\alias{buck.im}
\title{Bucks impute function}
\usage{
buck.im(
  model.pred = as.vector(model.pred),
  data = as.matrix(data),
  m = as.integer(m),
  n = as.integer(n)
)
}
\arguments{
\item{model.pred}{a vector, model prediction output from function model}

\item{data}{a matrix, datset from function miss.group}

\item{m}{an integer, variable to impute - enter as column number}

\item{n}{an integer, group number of interest}
}
\description{
This function imputes missing, NA vaules based on defined linear model
}
