% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BucksM.R
\name{miss.group}
\alias{miss.group}
\title{Calculates missing data groups and appends new column to original dataframe}
\usage{
miss.group(x)
}
\arguments{
\item{x}{a data frame}

\item{n}{number of rows}

\item{p}{number of columns, variables}

\item{r}{a binary matrix, 1 is mising value, 0 is complete}

\item{group}{missing pattern groups}

\item{new}{matrix with original dataframe and missing pattern group appened
and ordered as new column}
}
\description{
This function calculates missing data groups and appends new column
to original dataframe and orders it
}
