% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BucksM.R
\name{model}
\alias{model}
\title{Bucks complete case model & prediction}
\usage{
model(formula, data)
}
\arguments{
\item{formula}{a model to be fitted}

\item{data}{from miss.group function}

\item{m1}{model}
}
\description{
This function defines model based on complete cases and predicts
}
